refactor: 도메인에 맞게 이동
