+++ hyunique_app/src/main/java/com/kosa5/hyunique/product/mapper/ProductMapper.java
@@ -2,6 +2,7 @@ package com.kosa5.hyunique.product.mapper;
 
 import java.util.List;
 
+import com.kosa5.hyunique.post.vo.PostProductVO;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
@@ -13,4 +14,5 @@ public interface ProductMapper {
 
 	ProductDetailVO getProductDetailById(String productId);
 	List<PostThumbnailVO> getProductStyleById(@Param("productId") String productId,@Param("offset") int offset);
+	List<PostProductVO> selectSearchProductList(@Param("productName") String productName);
 }
