+++ hyunique_app/src/main/java/com/kosa5/hyunique/product/service/ProductService.java
@@ -2,6 +2,7 @@ package com.kosa5.hyunique.product.service;
 
 import java.util.List;
 
+import com.kosa5.hyunique.post.vo.PostProductVO;
 import com.kosa5.hyunique.post.vo.PostThumbnailVO;
 import com.kosa5.hyunique.product.vo.ProductDetailVO;
 
@@ -9,4 +10,6 @@ public interface ProductService {
 
 	ProductDetailVO getProductDetailById(String productId);
 	List<PostThumbnailVO> getProductStyleById(String productId, int offset);
+
+	List<PostProductVO> getSearchProductList(String productName);
 }
